+++ python/aistore/sdk/utils.py
@@ -5,7 +5,7 @@
 import logging
 import re
 from pathlib import Path
-from typing import Callable, Iterator, Optional, Tuple, Type, TypeVar
+from typing import Iterator, Optional, Tuple, Type, TypeVar
 from urllib.parse import urlparse
 
 import braceexpand
@@ -17,23 +17,11 @@ from msgspec import msgpack
 from pydantic.v1 import BaseModel, parse_raw_as
 
 from aistore.sdk.const import (
-    UTF_ENCODING,
     HEADER_CONTENT_TYPE,
     MSGPACK_CONTENT_TYPE,
     DEFAULT_LOG_FORMAT,
     XX_HASH_SEED,
 )
-from aistore.sdk.errors import (
-    AISError,
-    ErrBckNotFound,
-    ErrRemoteBckNotFound,
-    ErrBckAlreadyExists,
-    ErrObjNotFound,
-    ErrETLNotFound,
-    ErrETLAlreadyExists,
-    InvalidBckProvider,
-)
-from aistore.sdk.provider import Provider
 
 T = TypeVar("T")
 MASK = 0xFFFFFFFFFFFFFFFF  # 64-bit mask
@@ -50,122 +38,13 @@ class HttpError(BaseModel):
     Represents an error returned by the API.
     """
 
-    status: int
+    status_code: int
     message: str = ""
     method: str = ""
     url_path: str = ""
     remote_addr: str = ""
     caller: str = ""
     node: str = ""
-
-
-def decode_response_text(resp: requests.Response) -> str:
-    """
-    Convert the response body to a decoded string if it is bytes. Otherwise, return as is.
-
-    Args:
-        resp (requests.Response): The response object.
-
-    Returns:
-        str: The decoded response text (UTF-8 or ISO-8859-1).
-    """
-    error_text = resp.text
-    if isinstance(error_text, bytes):
-        try:
-            error_text = error_text.decode(UTF_ENCODING)
-        except UnicodeDecodeError:
-            error_text = error_text.decode("iso-8859-1")
-    return error_text
-
-
-def parse_http_error_or_raise(
-    resp: requests.Response, exc_class: Type[Exception]
-) -> HttpError:
-    """
-    Parse the response body as a HttpError object or raise a custom exception on failure.
-
-    Args:
-        resp (requests.Response): The response to parse.
-        exc_class (Type[Exception]): The exception class to raise if parsing fails.
-
-    Returns:
-        HttpError: The parsed HttpError model.
-
-    Raises:
-        exc_class: If the response text cannot be parsed into HttpError.
-    """
-    error_text = decode_response_text(resp)
-    try:
-        return parse_raw_as(HttpError, error_text)
-    except Exception as exc:
-        raise exc_class(resp.status_code, error_text) from exc
-
-
-def parse_ais_error(resp: requests.Response) -> AISError:
-    """
-    Parse raw text into an appropriate AISError object.
-
-    Args:
-        resp (requests.Response): The response from the AIS cluster.
-
-    Returns:
-        AISError: If the error doesn't match any specific conditions.
-        ErrBckNotFound: If the error message indicates a missing bucket.
-        ErrRemoteBckNotFound: If the error message indicates a missing remote bucket.
-        ErrObjNotFound: If the error message indicates a missing object.
-        ErrBckAlreadyExists: If the error message indicates a bucket already exists.
-        ErrETLAlreadyExists: If the error message indicates an ETL already exists
-        ErrETLNotFound: If the error message indicates a missing ETL.
-    """
-    exc_class = AISError
-    err = parse_http_error_or_raise(resp, exc_class)
-    status, message = err.status, err.message
-    prov, bck, obj = _extract_and_parse_url(message) or (None, None, None)
-    if prov:
-        try:
-            prov = Provider.parse(prov)
-        except InvalidBckProvider:
-            prov = None
-
-    if status == 404:
-        if "etl job" in message:
-            exc_class = ErrETLNotFound
-        elif prov:
-            if obj:
-                exc_class = ErrObjNotFound
-            elif prov.is_remote() or "@" in bck:
-                exc_class = ErrRemoteBckNotFound
-            else:
-                exc_class = ErrBckNotFound
-    elif status == 409:
-        if "etl job" in message:
-            exc_class = ErrETLAlreadyExists
-        elif prov and not obj:
-            exc_class = ErrBckAlreadyExists
-
-    return exc_class(status, message)
-
-
-# pylint: disable=unused-variable
-def handle_errors(
-    resp: requests.Response,
-    parse_error: Callable[[requests.Response], Exception],
-) -> None:
-    """
-    Error handling for requests made to the AIS Client
-
-    Args:
-        resp: requests.Response = Response received from the request
-        parse_error_fn: Function that processes error text and returns appropriate exceptions.
-    """
-    try:
-        # Will raise HTTPError if status code is 4xx or 5xx
-        resp.raise_for_status()
-    except requests.exceptions.HTTPError as http_err:
-        raise parse_error(resp) from http_err
-
-    # Final status code (post-redirection) wasn't 4xx/5xx. Could be 1xx or 3xx.
-    raise parse_error(resp)
 
 
 def probing_frequency(dur: int) -> float:
@@ -295,7 +174,15 @@ def parse_url(url: str) -> Tuple[str, str, str]:
     return parsed_url.scheme, parsed_url.netloc, path
 
 
-def _extract_and_parse_url(msg: str) -> Optional[Tuple[str, str, str]]:
+def extract_and_parse_url(msg: str) -> Optional[Tuple[str, str, str]]:
+    """
+    Parse any AIS info from a complete URL.
+    Args:
+        msg (str): Any string that may contain an AIS FQN
+
+    Returns:
+        Tuple[str, str, str]: Provider, bucket name, and object name, if contained in string, or None.
+    """
     url_pattern = r"[a-z0-9]+://[A-Za-z0-9@._/-]+"
     match = re.search(url_pattern, msg)
     if match: