+++ python/tests/unit/sdk/authn/test_authn_client.py
@@ -9,7 +9,6 @@ from urllib3 import Retry
 
 from aistore.sdk.authn import AuthNClient
 from aistore.sdk.authn.types import TokenMsg, LoginMsg
-from aistore.sdk.authn.utils import parse_authn_error
 
 from tests.utils import cases
 
@@ -22,14 +21,15 @@ class TestAuthNClient(unittest.TestCase):
 
     @patch("aistore.sdk.authn.authn_client.SessionManager")
     @patch("aistore.sdk.authn.authn_client.RequestClient")
-    def test_init_defaults(self, mock_request_client, mock_sm):
+    @patch("aistore.sdk.authn.authn_client.AuthNResponseHandler")
+    def test_init_defaults(self, mock_rh, mock_request_client, mock_sm):
         AuthNClient(self.endpoint)
         mock_request_client.assert_called_with(
             endpoint=self.endpoint,
             session_manager=mock_sm.return_value,
             timeout=None,
             token=None,
-            error_handler=parse_authn_error,
+            response_handler=mock_rh.return_value,
         )
 
     @cases(
@@ -40,7 +40,8 @@ class TestAuthNClient(unittest.TestCase):
     )
     @patch("aistore.sdk.authn.authn_client.SessionManager")
     @patch("aistore.sdk.authn.authn_client.RequestClient")
-    def test_init(self, test_case, mock_request_client, mock_sm):
+    @patch("aistore.sdk.authn.authn_client.AuthNResponseHandler")
+    def test_init(self, test_case, mock_rh, mock_request_client, mock_sm):
         skip_verify, ca_cert, timeout, retry, token = test_case
         # print all vars
         print(
@@ -64,7 +65,7 @@ class TestAuthNClient(unittest.TestCase):
             session_manager=mock_sm.return_value,
             timeout=timeout,
             token=token,
-            error_handler=parse_authn_error,
+            response_handler=mock_rh.return_value,
         )
 
     @patch("aistore.sdk.request_client.RequestClient.request_deserialize")