+++ python/tests/unit/sdk/test_utils.py
@@ -1,31 +1,18 @@
 #
 # Copyright (c) 2022-2025, NVIDIA CORPORATION. All rights reserved.
 #
-
-import json
 import unittest
 from unittest.mock import Mock, patch, mock_open
 
 from msgspec import msgpack
 from requests import Response
 
-from aistore.sdk.authn.utils import parse_authn_error
 from aistore.sdk.const import MSGPACK_CONTENT_TYPE, HEADER_CONTENT_TYPE, XX_HASH_SEED
-from aistore.sdk.errors import (
-    AISError,
-    ErrRemoteBckNotFound,
-    ErrBckNotFound,
-    ErrBckAlreadyExists,
-    ErrObjNotFound,
-    ErrETLAlreadyExists,
-    ErrETLNotFound,
-)
+
 from aistore.sdk.utils import (
     decode_response,
     expand_braces,
     get_file_size,
-    handle_errors,
-    parse_ais_error,
     probing_frequency,
     read_file_bytes,
     validate_directory,
@@ -33,88 +20,12 @@ from aistore.sdk.utils import (
     xoshiro256_hash,
     get_digest,
 )
-from aistore.sdk.authn.errors import (
-    AuthNError,
-    ErrUserNotFound,
-    ErrUserAlreadyExists,
-    ErrRoleNotFound,
-    ErrRoleAlreadyExists,
-    ErrClusterNotFound,
-    ErrClusterAlreadyRegistered,
-    ErrUserInvalidCredentials,
-)
 from tests.const import PREFIX_NAME
-from tests.utils import cases, case_matrix
+from tests.utils import cases
 
 
 # pylint: disable=unused-variable
 class TestUtils(unittest.TestCase):
-    def _get_err_parse_fn(self, err_type):
-        if issubclass(err_type, AuthNError):
-            return parse_authn_error
-        return parse_ais_error
-
-    @cases(AISError, AuthNError)
-    def test_handle_error_no_text(self, err_type):
-        mock_response = Mock(text="")
-        with self.assertRaises(err_type):
-            handle_errors(mock_response, self._get_err_parse_fn(err_type))
-        mock_response.raise_for_status.assert_called()
-
-    @cases(AISError, AuthNError)
-    def test_handle_error_decode_err(self, err_type):
-        err_status = 300
-        err_msg = "error message iso-8859-1"
-        expected_text = json.dumps({"status": err_status, "message": err_msg})
-        # Fail initial decoding, then return the decoded text
-        decode_err = UnicodeDecodeError("1", b"2", 3, 4, "5")
-        mock_iso_text = Mock(spec=bytes)
-        mock_iso_text.decode.side_effect = [decode_err, expected_text]
-        self.handle_err_exec_assert(err_type, err_status, err_msg, mock_iso_text)
-
-    @case_matrix([399, 500], [AISError, AuthNError])
-    def test_handle_error_base_errors(self, err_status, err_type):
-        err_msg = "error message"
-        expected_text = json.dumps({"status": err_status, "message": err_msg})
-        mock_text = Mock(spec=bytes)
-        mock_text.decode.return_value = expected_text
-        self.handle_err_exec_assert(err_type, err_status, err_msg, mock_text)
-
-    @cases(
-        ('aws bucket "s3://test-bck" does not exist', ErrRemoteBckNotFound, 404),
-        ('remote bucket "ais://@test-bck" does not exist', ErrRemoteBckNotFound, 404),
-        ('bucket "ais://test-bck" does not exist', ErrBckNotFound, 404),
-        ('bucket "ais://test-bck" already exists', ErrBckAlreadyExists, 409),
-        ("ais://test-bck/test-obj does not exist", ErrObjNotFound, 404),
-        ("etl job test-etl-job already exists", ErrETLAlreadyExists, 409),
-        ("etl job test-etl-job does not exist", ErrETLNotFound, 404),
-        ('user "test-user" does not exist', ErrUserNotFound, 404),
-        ('user "test-user" already exists', ErrUserAlreadyExists, 409),
-        ('role "test-role" does not exist', ErrRoleNotFound, 404),
-        ('role "test-role" already exists', ErrRoleAlreadyExists, 409),
-        ("cluster test-cluster does not exist", ErrClusterNotFound, 404),
-        (
-            "cluster OnBejJEpe[OnBejJEpe] already registered",
-            ErrClusterAlreadyRegistered,
-            409,
-        ),
-        ("invalid credentials", ErrUserInvalidCredentials, 401),
-    )
-    def test_handle_errors(self, test_case):
-        err_msg, expected_err, err_status = test_case
-        expected_text = json.dumps({"status": err_status, "message": err_msg})
-        mock_text = Mock(spec=bytes)
-        mock_text.decode.return_value = expected_text
-        self.handle_err_exec_assert(expected_err, err_status, err_msg, mock_text)
-
-    def handle_err_exec_assert(self, err_type, err_status, err_msg, mock_err_text):
-        err_parse_fn = self._get_err_parse_fn(err_type)
-        mock_response = Mock(text=mock_err_text)
-        with self.assertRaises(err_type) as context:
-            handle_errors(mock_response, err_parse_fn)
-        self.assertEqual(err_msg, context.exception.message)
-        self.assertEqual(err_status, context.exception.status_code)
-
     @cases((0, 0.1), (-1, 0.1), (64, 1), (128, 2), (100000, 1562.5))
     def test_probing_frequency(self, test_case):
         self.assertEqual(test_case[1], probing_frequency(test_case[0]))